Fix dedup function return type — group_size BIGINT (was INT)
COUNT(*) returns BIGINT in Postgres; declaring group_size as INT caused
"structure of query does not match function result type" on every call.
One-line fix.

+++ migrations/085_dedup_properties_function.sql
@@ -43,7 +43,7 @@ RETURNS TABLE (
   normalized_address    TEXT,
   survivor_id           UUID,
   merged_ids            UUID[],
-  group_size            INT,
+  group_size            BIGINT,
   fk_relinks_summary    JSONB,
   action                TEXT
 )
